Implement BadArgs in new refactor
