+++ cogs/canvas.py
@@ -10,6 +10,7 @@ from discord.ext import commands
 from dotenv import load_dotenv
 
 from handlers.canvas_handler import CanvasHandler
+from cogs.meta import BadArgs
 
 CANVAS_COLOR = 0xe13f2b
 CANVAS_THUMBNAIL_URL = "https://lh3.googleusercontent.com/2_M-EEPXb2xTMQSTZpSUefHR3TjgOCsawM3pjVG47jI-BrHoXGhKBpdEHeLElT95060B=s180"
@@ -32,7 +33,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         c_handler.track_course(course_ids)
 
@@ -53,7 +54,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         c_handler.untrack_course(course_ids)
 
@@ -83,7 +84,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         if args and args[0].startswith("-due"):
             due = args[1]
@@ -116,7 +117,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         if ctx.message.channel not in c_handler.live_channels:
             c_handler.live_channels.append(ctx.message.channel)
@@ -135,7 +136,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         if ctx.message.channel in c_handler.live_channels:
             c_handler.live_channels.remove(ctx.message.channel)
@@ -168,7 +169,7 @@ class Canvas(commands.Cog):
         c_handler = self._get_canvas_handler(ctx.message.guild)
 
         if not isinstance(c_handler, CanvasHandler):
-            return await ctx.send("Canvas Handler doesn't exist.", delete_after=5)
+            raise BadArgs("Canvas Handler doesn't exist.", False, ctx.command)
 
         if args and args[0].startswith("-since"):
             since = args[1]
